Finish setup refactor
